% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brom.R
\name{brom}
\alias{brom}
\title{Bivariate Record Occurrence Model Fitting}
\usage{
brom(
  formula,
  data,
  coords,
  extra.coords,
  sp.A = FALSE,
  matrix.mean.A = matrix(1, nrow = nrow(coords)),
  site,
  year,
  day,
  scale.cov = TRUE,
  decay.prior = c("fixed", "unif", "gamma"),
  inits = NULL,
  prior = list(beta = 0.01, a = 1/5^2, decay = c(2, 100), sigma = c(2, 1), decay_A =
    0.01),
  n.sims = 1000,
  n.thin = 1,
  n.burnin = 1000,
  n.report = 100
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the
model to be fitted. (Has some requirements with \code{trend},
\code{lag.tx}, and \code{lag.tn}).}

\item{data}{a data frame (\eqn{N \times p}) containing the variables in the
model.}

\item{coords}{a matrix or data frame (\eqn{n \times 2}) containing the
coordinates of the sites.}

\item{extra.coords}{a matrix or data frame (\eqn{n \times k}) containing
spatial covariates to be included in the spatial correlation function.}

\item{sp.A}{a boolean indicating a spatially-varying coregionalization.}

\item{matrix.mean.A}{a matrix (\eqn{n \times q}) containing the variables
in the mean of A (if sp.A is TRUE).}

\item{site, year, day}{vectors (\eqn{N \times 1}) containing the site, year,
and day the measurements were taken.}

\item{scale.cov}{a boolean indicating whether or not the columns of the
design matrix must be scaled to have zero mean and unit variance.}

\item{decay.prior}{a character string indicating the prior for the decay
parameter, gamma, uniform in an interval, or fixed.}

\item{inits}{a vector of initial values (if NULL, random values).}

\item{prior}{a list containing the parameters of the prior distributions:
the precision for beta; the scale for a; for the main decay if fixed the
fixed values, 1 + 1 for each column in \code{extra.coords}, if gamma(a, b)
first all a's then all b's (again 1 + 1 each), if unif(a, b) the same;
both hyperparameters of the gamma priors, the fixed decay for the
spatially-varying a's.}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations not discarded.
(ii) Thinning rate.
(iii) Number of iterations discarded at the beginning.
(iv) Report the number of iterations rate.}
}
\value{
A \code{brom} list with many elements. Samples from the model
  parameters are in \code{params} where rows are MCMC simulations and cols
  are parameters.
}
\description{
This function fits the Bivariate Record Occurrence Model in a Bayesian
  framework using MCMC.
}
\details{
The fitting algorithm is a data-augmentation Metropolis-within-Gibbs MCMC
  algorithm.

  \code{formula} requires two variables, \code{lag.tx} and \code{lag.tn},
  that are the record indicators of the previous day for max and min.

  \code{data} requires column names \code{trend} (annual trend),
  \code{lag.tx}, and \code{lag.tn}.

  The response variable must be \code{Y}. It can be 0's and 1's indicating
  the occurrence of a record, but values \eqn{1/r} (\eqn{r=2,3,\ldots}) are
  allowed indicating the probability of a tied record being a strong record.
  (If all values are 0's and 1's the function could return an error.)
}
\note{
It is necessary to have data ordered as using if c() over a
  3-dimensional array where the first dimension is year, the second is day,
  and the third is site. This is, first all data for the first site, then
  for the first day and it progresses by years. (This will also be necessary
  for prediction.)
}
\examples{
# brom(cbind(tx, tn) ~ trend + lag.tx + lag.tn, data, coords)

}
