% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCconvergence.R
\name{MCMCconvergence}
\alias{MCMCconvergence}
\title{Convergence diagnostics for the models}
\usage{
MCMCconvergence(object)
}
\arguments{
\item{object}{A matrix of models (same model different chains) by columns or
a list with each list a chain and each element of the chain several models
by columns associated with different folds.}
}
\value{
Diagnostics matrix.
}
\description{
This function is used to check the convergence of the models.
}
\author{
Jorge Castillo-Mateo
}
