% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.brom}
\alias{predict.brom}
\title{Predict Method for Bivariate Record Occurrence Model Fits}
\usage{
\method{predict}{brom}(
  object,
  newdata,
  newcoords,
  newextra.coords,
  newsite,
  newyear,
  newday,
  type = c("KFCV", "sample")
)
}
\arguments{
\item{object}{Object of class inheriting from "brom"}

\item{newdata}{An optional list with \code{x} in which to look for variables
with which to predict, \code{coords} and \code{site} in which to predict,
and \code{day} or \code{year}.
If omitted, the fitted values are used.}

\item{newcoords, newextra.coords}{New coordinates for the sites in
\code{newdata$site} or \code{newsite}.}

\item{newsite, newyear, newday}{A vector where for each observation includes
the site, year or day.}

\item{type}{the type of prediction required. For \eqn{K}-fold
cross-validation \code{"KFCV"} uses observed indicators from \code{data}
and samples weak records from their prior probability.}
}
\value{
A matrix where rows are MCMC posterior predictive samples and cols
  are observations.
}
\description{
Predicted values based on \code{brom} (bivariate record occurrence model)
  object.
}
\details{
Predicted values are on the scale of the response (probabilities).
}
