% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbrom.R
\docType{package}
\name{spbrom-package}
\alias{spbrom-package}
\alias{spbrom}
\title{spbrom: Spatial Bivariate Record Occurrence Model}
\description{
Fitting, prediction and validation of bivariate spatial models for the
  occurrence of record-breaking events over time in two dependent processes.
}
\author{
\packageAuthor{spbrom}

  Maintainer: \packageMaintainer{spbrom}
}
