% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics for Validation}
\usage{
metrics(Y, p, year, metric = c("DIC", "CV"), D1 = 2:31, D2 = 32:63)
}
\arguments{
\item{Y}{The observed values (vector)}

\item{p}{The estimated probabilities (matrix with replicates by rows)}

\item{year}{The year associated with the values (vector)}

\item{metric}{A character string \code{"DIC"} or \code{"CV"}}

\item{D1, D2}{Two shorter year periods on which to calculate the metrics, in
addition to the period of all years}
}
\value{
If \code{metric = "CV"}; the metrics (vector):
  \item{AUC}{Area under the ROC curve}
  \item{AUC1}{AUC in \code{D1}}
  \item{AUC2}{AUC in \code{D2}}
  \item{J}{Jaccard index}
  \item{J1}{J in \code{D1}}
  \item{J2}{J in \code{D2}}

  If \code{metric = "DIC"}; a vector with DIC, D (deviance), and penalty (pD)
}
\description{
This function is used to validate the predicted probabilities of record
  with the actual values.
}
\details{
Values different from 0 or 1 (tied records) are removed from the metrics.
}
\author{
Jorge Castillo-Mateo
}
